requirements.txt
yfinance>=0.2.40
pandas>=2.2.3
numpy>=1.26.4
scikit-learn>=1.5.0
vaderSentiment==3.3.2
requests>=2.32.3
schedule>=1.2.2
streamlit>=1.35.0
plotly>=5.22.0
python-dotenv>=1.0.1
openpyxl>=3.1.4
